refactor: split ResearchArtifact monolith into 14-file module
