+++ src/components/DeepResearch/ResearchArtifact/index.ts
@@ -1,0 +1,2 @@
+export { ResearchArtifact } from './ResearchArtifact';
+export type { ResearchArtifactProps } from './types';
