+++ src/components/primitives/index.ts
@@ -1,4 +1,5 @@
 export { Card } from './Card';
 export { Stack } from './Stack';
 export { Row } from './Row';
+export { Label } from './Label';
 export { EmptyState } from './EmptyState';
